BookFlow_API/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.2.17
filelock==3.16.0
platformdirs==4.3.2
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.26.4
Django>=4.0,<5.0
djangorestframework==3.14.0
django-filter